fix: using environment variables for config
Adjusted some files to use environment variables for configuration.

Also renamed files to better represent there contents.

For instance, example.ini -> example.env because the ini
format is not being used at the top level anymore.

Prefixed the scripts with s_ so they would be grouped together in
directory.

$CONTAINER_RUNTIME variable is now used to determine whether docker or
podman is in use.

README.md is updated with instructions for using environment variables.